add small_pprof

+++ init_tools/clean.sh
@@ -1,1 +1,6 @@
+path=`find . -maxdepth 1 -name clean.sh`
+if [[ $path != "" ]];then
+    ./clean.sh
+    exit 0
+fi
 rm -rf build
